SrvrMon-main/requirements.txt
streamlit>=1.31.0
pandas>=2.0.0
psutil>=5.8
docker>=5.0.0
python-dateutil>=2.8.0
PyYAML>=5.0